Extract Duke's Ui to its own class
